Switched to using Core Components bundle version for both header and footer.
