+++ examples/ui.apps/src/content/jcr_root/apps/core-components-examples/components/version/footer/footer.html
@@ -13,5 +13,4 @@
   ~ See the License for the specific language governing permissions and
   ~ limitations under the License.
   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/-->
-<p data-sly-use.version="${'PackageVersionHelper' @ packageName='core.wcm.components.content'}"
->Built with <a href="https://github.com/adobe/aem-core-wcm-components/releases/">Core Components ${version.packageVersion}</a>.</p>
+<p data-sly-use.version="VersionHelper">Built with <a href="https://github.com/adobe/aem-core-wcm-components/releases/">Core Components ${version.bundleVersion}</a>.</p>
